requirements.txt
mashumaro==3.12
pytest-homeassistant-custom-component==0.13.100
mypy==1.4.1
pre-commit==3.6.2
ruff==0.0.292
pytest-asyncio==0.23.4
